+++ Pod/Classes/ViewControllers/ARComponentViewController.h
@@ -5,9 +5,23 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface ARComponentViewController : UIViewController
+/// A way of letting consumers of Emission know that this View Controller
+/// needs to be treated a little differently with positioning
+@protocol ARComponentFullBleedViewSizing <NSObject>
 
+@property (nonatomic, readonly) BOOL fullBleed;
+
+@end
+
+
+/// The base subclass for any UIViewController that wraps a React Native root view
+@interface ARComponentViewController : UIViewController <ARComponentFullBleedViewSizing>
+
+/// The React View
 @property (nonatomic, strong) RCTRootView *rootView;
+
+// If set to true, the view controller will start under th status bar (iPhone X, iPad Pro, etc...)
+@property (nonatomic, assign) BOOL fullBleed;
 
 - (instancetype)init NS_UNAVAILABLE;
 - (instancetype)initWithNibName:(nullable NSString *)nibNameOrNil bundle:(nullable NSBundle *)nibBundleOrNil NS_UNAVAILABLE;