+++ Pod/Classes/ViewControllers/ARFairComponentViewController.m
@@ -24,8 +24,11 @@
 
 - (instancetype)initWithFairID:(NSString *)fairID emission:(nullable AREmission *)emission
 {
-    return [super initWithEmission:emission moduleName:@"Fair" initialProperties:@{ @"fairID": fairID }];
+    self = [super initWithEmission:emission moduleName:@"Fair" initialProperties:@{ @"fairID": fairID }];
+    if (self) {
+        self.fullBleed = YES;
+    }
+    return self;
 }
-
 
 @end