+++ CHANGELOG.md
@@ -67,6 +67,10 @@
 - Adds Shows to Global Saves and Follows view -ashley
 - Adds separator between fair entities - kieran
 - Adds Fair analytics - kieran
+- Refactor and style new filter header in map view - luc
+- Move filters down to city view - luc
+- Adds new map filter design to Map view - luc
+- Support full bleed view controllers - luc
 
 ### 1.8.2